Move styles and types into own files for Common Components
