+++ components/Common/AddExercisePopUp/ExerciseList/ExerciseList.tsx
@@ -1,16 +1,10 @@
 import { FlatList } from "react-native";
+
 import { lightHaptic } from "@/utils/haptics/haptics";
-import { Exercise } from "@/db/schema";
-import { ExerciseInput } from "@/app/(tabs)/_layout";
+
+import { ExerciseListProps } from "./types";
 
 import ExerciseItem from "./ExerciseItem";
-
-interface ExerciseListProps {
-    data: Exercise[];
-    selectedExercises: Map<number, ExerciseInput>;
-    addExercise: (exercise: ExerciseInput) => void;
-    removeExercise: (id: number) => void; 
-}
 
 export default function ExerciseList({ data, selectedExercises, addExercise, removeExercise }: ExerciseListProps) {
     return (
