+++ components/Common/AddExercisePopUp/ExerciseList/styles.ts
@@ -1,0 +1,18 @@
+import { StyleSheet } from "react-native";
+
+export const styles = StyleSheet.create({
+    itemContainer: {
+        borderBottomWidth: 1,
+        borderColor: "rgba(255, 255, 255, .8)",
+        padding: 15,
+    },
+    itemText: {
+        color: "white",
+        fontSize: 18,
+        fontWeight: "700"
+    },
+    schema: {
+        color: "rgba(166, 166, 166, .8)",
+        fontWeight: "700"
+    }
+});
