+++ components/Common/AddExercisePopUp/Filters.tsx
@@ -1,13 +1,10 @@
-import { StyleSheet, TextInput, TouchableOpacity, View } from "react-native";
+import { TextInput, TouchableOpacity, View } from "react-native";
 import FontAwesome6 from '@expo/vector-icons/FontAwesome6';
 
 import { lightHaptic } from "@/utils/haptics/haptics";
 
-interface FiltersProps {
-    searchValue: string;
-    setSearchValue: React.Dispatch<React.SetStateAction<string>>;
-    openCreateExerciseForm: () => void;
-}
+import { styles } from "./styles";
+import { FiltersProps } from "./types";
 
 export default function Filters({ searchValue, setSearchValue, openCreateExerciseForm }: FiltersProps) {
     return (
@@ -27,42 +24,10 @@ export default function Filters({ searchValue, setSearchValue, openCreateExercis
                         lightHaptic();
                         openCreateExerciseForm();
                     }}
-                    style={styles.addButtonWrapper}
+                    style={styles.filterAddButtonWrapper}
                 >
                 <FontAwesome6 name="add" size={24} color="white" />
             </TouchableOpacity>
         </View>
     )
 }
-
-const styles = StyleSheet.create({
-    container: {
-        alignItems: "center",
-        backgroundColor: "rgba(210, 210, 210, 1)",
-        flexDirection: "row",
-        justifyContent: "center",
-        paddingHorizontal: 15,
-        paddingVertical: 15,
-    },
-    textInputWrapper: {
-        alignItems: "center",
-        justifyContent: "center",
-        paddingRight: 10,
-        flex: 1
-    },
-    textInput: {
-        backgroundColor: "rgba(0, 0, 0, .5)",
-        borderRadius: 10,
-        color: "white",
-        fontSize: 18,
-        fontWeight: "700",
-        paddingHorizontal: 12,
-        paddingVertical: 7,
-        width: "100%"
-    },
-    addButtonWrapper: {
-        backgroundColor: "rgba(0, 0, 0, .5)",
-        borderRadius: 8,
-        padding: 5
-    }
-})
