+++ components/Common/AddExercisePopUp/Header.tsx
@@ -1,13 +1,10 @@
-import { StyleSheet, Text, TouchableOpacity, View } from "react-native";
+import { Text, TouchableOpacity, View } from "react-native";
 import Feather from '@expo/vector-icons/Feather';
 
 import { lightHaptic } from "@/utils/haptics/haptics";
 
-interface HeaderProps {
-    close: () => void;
-    selectedExerciseCount: number;
-    onAdd: () => void;
-}
+import { styles } from "./styles";
+import { HeaderProps } from "./types";
 
 export default function Header({ close, selectedExerciseCount, onAdd }: HeaderProps) {
     let addButton;
@@ -55,45 +52,3 @@ export default function Header({ close, selectedExerciseCount, onAdd }: HeaderPr
         </View>
     )
 }
-
-const styles = StyleSheet.create({
-    header: {
-        alignItems: "center",
-        flexDirection: "row",
-        height: 50,
-        paddingHorizontal: 15,
-        width: "100%"
-    },
-    exitButtonWrapper: {
-        alignItems: "center",
-        backgroundColor: "rgba(255, 255, 255, .3)",
-        borderRadius: 5,
-        justifyContent: "center",
-        padding: 3
-    },
-    headerTextWrapper: {
-        alignItems: "center",
-        flex: 2,
-    },
-    headerText: {
-        color: "white",
-        fontSize: 20,
-        fontWeight: "800",
-    },
-    addButtonWrapper: {
-        width: 70
-    },
-    addButtonText: {
-        color: "rgba(250, 250, 250, .3)",
-        fontSize: 18,
-        fontWeight: "700"
-    },
-    exitWrapper: {
-        alignItems: "flex-start",
-        flex: 1
-    },
-    addWrapper: {
-        alignItems: "flex-end",
-        flex: 1
-    }
-})
