+++ components/Common/Buttons/AddSetButton/styles.ts
@@ -1,0 +1,15 @@
+import { StyleSheet } from "react-native";
+
+export const styles = StyleSheet.create({
+    container: {
+        flexDirection: "row",
+        paddingHorizontal: 15,
+        paddingVertical: 5
+    },
+    buttonWrapper: {
+        backgroundColor: "rgba(0, 0, 0, .4)",
+        borderRadius: 8,
+        flexDirection: "row",
+        padding: 5
+    }
+});
