+++ components/Common/Buttons/AddSetButton/types.ts
@@ -1,0 +1,3 @@
+export interface AddSetButtonProps {
+    addSet: () => void;
+}
