+++ components/Common/Buttons/ExitButton/styles.ts
@@ -1,0 +1,9 @@
+import { StyleSheet } from "react-native";
+
+export const styles = StyleSheet.create({
+    container: {
+        backgroundColor: "rgba(255, 255, 255, .3)",
+        borderRadius: 5,
+        justifyContent: "center"
+    }
+});
