+++ components/Common/Buttons/ExitButton/types.ts
@@ -1,0 +1,3 @@
+export interface ExitButtonProps {
+    onPress: () => void;
+}
