+++ components/Common/CompletedWorkoutCard/styles.ts
@@ -1,0 +1,24 @@
+import { StyleSheet } from "react-native";
+
+export const styles = StyleSheet.create({
+    cardContainer: {
+        borderWidth: 1,
+        borderColor: "white",
+        borderRadius: 10,
+        padding: 15
+    },
+    cardHeader: {
+        color: "white",
+        fontSize: 24,
+        fontWeight: "800",
+        marginBottom: 10
+    },
+    container: {
+        paddingTop: 5
+    },
+    exercises: {
+        color: "white",
+        fontSize: 16,
+        fontWeight: "700"
+    }
+});
