+++ components/Common/ConfirmationPopUp/types.ts
@@ -1,0 +1,12 @@
+export interface ConfirmationHeaderProps {
+    header: string;
+    closeConfirmation: () => void;
+}
+
+export interface ConfirmationPopUpProps {
+    header: string;
+    description: string;
+    showConfirmation: boolean;
+    closeConfirmation: () => void;
+    onConfirm: () => void;
+}
