+++ components/Common/Duration/styles.ts
@@ -1,0 +1,17 @@
+import { StyleSheet } from "react-native";
+
+export const styles = StyleSheet.create({
+    container: {
+        alignItems: "center",
+        flexDirection: "row",
+        paddingBottom: 10
+    },
+    iconWrapper: {
+        paddingRight: 5,
+    },
+    duration: {
+        color: "white",
+        fontSize: 16,
+        fontWeight: "600"
+    }
+});
