+++ components/Common/Duration/types.ts
@@ -1,0 +1,3 @@
+export interface DurationProps {
+    duration: number;
+}
