+++ components/Common/ExerciseDetailViewer/ExerciseDetail/ExerciseDescription.tsx
@@ -1,19 +1,16 @@
-import { StyleSheet, Text, View } from "react-native";
+import { Text, View } from "react-native";
 
-import { Exercise } from "@/db/schema";
-
-interface ExerciseDescriptionProps {
-    exercise?: Exercise;
-}
+import { styles } from "./styles";
+import { ExerciseDescriptionProps } from "./types";
 
 export default function ExerciseDescription({ exercise }: ExerciseDescriptionProps) {
     return (
-        <View style={styles.container}>
+        <View style={styles.descriptionContainer}>
             {exercise && (
                 <>
-                    <View style={styles.schemaWrapper}>
-                        <Text style={styles.schemaLabel}>Schema</Text>
-                        <Text style={styles.schemaValue}>{exercise.schema}</Text>
+                    <View style={styles.descriptionSchemaWrapper}>
+                        <Text style={styles.descriptionSchemaLabel}>Schema</Text>
+                        <Text style={styles.descriptionSchemaValue}>{exercise.schema}</Text>
                     </View>
                     <View>
                         <Text style={styles.descriptionLabel}>Description</Text>
@@ -24,35 +21,3 @@ export default function ExerciseDescription({ exercise }: ExerciseDescriptionPro
         </View>
     )
 }
-
-const styles = StyleSheet.create({
-    container: {
-        flex: 1,
-        padding: 15
-    },
-    schemaWrapper: {
-        paddingBottom: 12
-    },
-    schemaLabel: {
-        color: "white",
-        fontSize: 24,
-        fontWeight: "700",
-        marginRight: 10
-    },
-    schemaValue: {
-        color: "rgba(200, 200, 200, 1)",
-        fontSize: 16,
-        fontWeight: "600"
-    },
-    descriptionLabel: {
-        color: "white",
-        fontSize: 24,
-        fontWeight: "700",
-        marginRight: 10
-    },
-    descriptionValue: {
-        color: "rgba(200, 200, 200, 1)",
-        fontSize: 16,
-        fontWeight: "600"
-    }
-})
