+++ components/Common/ExerciseDetailViewer/ExerciseDetail/ExerciseDetail.tsx
@@ -4,22 +4,14 @@ import { useQuery } from "@tanstack/react-query";
 
 import { getExercise } from "@/db/queries";
 
+import { styles } from "./styles";
+import { ActiveTab, ExerciseDetailProps } from "./types";
+
 import TabNavigator from "./utils/TabNavigator/TabNavigator";
 import ExerciseAnalytics from "./ExerciseAnalytics";
 import ExerciseHistory from "./ExerciseHistory";
 import ExerciseDescription from "./ExerciseDescription";
 import EditExerciseForm from "@/components/Forms/EditExercise/EditExerciseForm";
-
-export type ActiveTab = "Analytics" | "History" | "Description";
-
-interface ExerciseDetailProps {
-    showDetails: boolean;
-    closeDetails: () => void;
-    showForm: boolean;
-    openForm: () => void;
-    closeForm: () => void;
-    id: number;
-}
 
 export default function ExerciseDetail({ id, showDetails, closeDetails, showForm, openForm, closeForm }: ExerciseDetailProps) {
     const [activeTab, setActiveTab] = useState<ActiveTab>("Description");
@@ -46,9 +38,9 @@ export default function ExerciseDetail({ id, showDetails, closeDetails, showForm
         >
             <View style={styles.transparentBackground}>
                 <View style={styles.detailsBackground}>
-                    <View style={styles.header}>
+                    <View style={styles.detailsHeader}>
                         <Button title="Exit" onPress={closeDetails} />
-                        <Text style={styles.headerText}>
+                        <Text style={styles.detailsHeaderText}>
                             {data?.length == 1 ? data[0].name : "Exercise"} Details
                         </Text>
                         <Button title="Edit" onPress={openForm} />
@@ -68,35 +60,3 @@ export default function ExerciseDetail({ id, showDetails, closeDetails, showForm
         </>
     )
 }
-
-const styles = StyleSheet.create({
-    transparentBackground: {
-        alignItems: "center",
-        backgroundColor: "rgba(0, 0, 0, .8)",
-        flex: 1,
-        justifyContent: "center"
-    },
-    detailsBackground: {
-        aspectRatio: 5/9,
-        backgroundColor: "rgba(80, 80, 80, 1)",
-        borderRadius: 15,
-        width: "95%"
-    },
-    header: {
-        alignItems: "center",
-        flexDirection: "row",
-        justifyContent: "space-between",
-        paddingHorizontal: 15,
-        paddingTop: 15
-    },
-    headerText: {
-        color: "white",
-        fontSize: 18,
-        fontWeight: "700"
-    },
-    navigator: {
-        alignItems: "center",
-        justifyContent: "center",
-        padding: 15
-    }
-})
