+++ components/Common/Timestamp/styles.ts
@@ -1,0 +1,19 @@
+import { StyleSheet } from "react-native";
+
+export const styles = StyleSheet.create({
+    container: {
+        flexDirection: "row",
+        marginBottom: 5
+    },
+    iconWrapper: {
+        paddingRight: 5
+    },
+    timeStampWrapper: {
+        justifyContent: "flex-end"
+    },
+    date: {
+        color: "white",
+        fontSize: 18,
+        fontWeight: "600"
+    }
+})
