+++ components/Common/Timestamp/types.ts
@@ -1,0 +1,3 @@
+export interface TimestampProps {
+    timestamp: number;
+}
